Fixed Styling and created new styling files
