+++ client/src/pages/Dashboard/styles/styledMenu.js
@@ -1,0 +1,25 @@
+import { withStyles } from "@material-ui/styles";
+import React from "react";
+import { Menu } from "@material-ui/core";
+
+const StyledMenu = withStyles({
+  paper: {
+    border: "1px solid #d3d4d5"
+  }
+})(props => (
+  <Menu
+    elevation={0}
+    getContentAnchorEl={null}
+    anchorOrigin={{
+      vertical: "bottom",
+      horizontal: "center"
+    }}
+    transformOrigin={{
+      vertical: "top",
+      horizontal: "center"
+    }}
+    {...props}
+  />
+));
+
+export default StyledMenu;
